Server/webplagiarism/requirements.txt
flask~=1.1.2
flask-wtf
WTForms~=2.3.3
email-validator
flask-sqlalchemy
flask-bcrypt
flask-login
Pillow
flask-mail
numpy==1.19.4
textdistance
chars2vec
scikit-learn==0.24.0
sklearn
music21==6.5.0
Keras==2.4.3
tensorflow==2.4.0
tensorflow-estimator==2.4.0
waitress
